+++ src/Java/myservicebus-abstractions/src/main/java/com/myservicebus/ReceiveTransport.java
@@ -1,7 +1,13 @@
 package com.myservicebus;
 
+import java.time.Duration;
+
 public interface ReceiveTransport {
     void start() throws Exception;
 
     void stop() throws Exception;
+
+    default void stop(Duration timeout) throws Exception {
+        stop();
+    }
 }